Change vector types
For every vector operation, there is a choice of saying that
it returns a type like __m128i (Intel's name) or of
giving a more informative/structured type like u32x4 indicating that
it returns 4 32-bit values.

For LLVM purposes, we _must_ use the structured types.
For MIR purposes, we probably need to use the Intel types.

This change switches from Intel types to structured types.
The intention is to implement a shim layer for MIR purposes that
converts arguments/results as needed and calls the LLVM-friendly version